requirements.txt
asgiref==3.7.2
Django==5.0.3
django-jazzmin==2.6.0
djangorestframework==3.15.0
psycopg2==2.9.9
python-dotenv==1.0.1
sqlparse==0.4.4
typing_extensions==4.10.0
